+++ paddle/cinn/auto_schedule/search_space/auto_gen_rule/multi_level_tiling.cc
@@ -30,10 +30,10 @@
 #include "paddle/cinn/ir/collect_ir_nodes.h"
 #include "paddle/cinn/ir/ir.h"
 #include "paddle/cinn/ir/ir_base.h"
+#include "paddle/cinn/ir/ir_copy.h"
 #include "paddle/cinn/ir/ir_printer.h"
 #include "paddle/cinn/ir/ir_schedule.h"
 #include "paddle/cinn/ir/tensor.h"
-#include "paddle/cinn/optim/ir_copy.h"
 
 namespace cinn {
 namespace auto_schedule {
