+++ paddle/cinn/auto_schedule/search_space/auto_gen_rule/skip_rule.cc
@@ -18,8 +18,8 @@
 
 #include "paddle/cinn/auto_schedule/search_space/auto_gen_rule/auto_gen_rule.h"
 #include "paddle/cinn/common/target.h"
+#include "paddle/cinn/ir/ir_copy.h"
 #include "paddle/cinn/ir/ir_schedule.h"
-#include "paddle/cinn/optim/ir_copy.h"
 
 namespace cinn {
 namespace auto_schedule {
