+++ paddle/cinn/auto_schedule/task/task_registry.h
@@ -19,8 +19,8 @@
 #include <mutex>
 #include <string>
 
+#include "paddle/cinn/ir/ir_copy.h"
 #include "paddle/cinn/ir/ir_schedule.h"
-#include "paddle/cinn/optim/ir_copy.h"
 #include "paddle/cinn/utils/registry.h"
 
 namespace cinn {
