+++ paddle/cinn/ir/CMakeLists.txt
@@ -27,7 +27,8 @@ gather_srcs(
   intrinsic_ops.cc
   layout.cc
   schedule_desc.cc
-  ir_compare.cc)
+  ir_compare.cc
+  ir_copy.cc)
 
 # cinn_cc_test(test_ir SRCS ir_test.cc DEPS core)
 # cinn_cc_test(test_ir_printer SRCS ir_printer_test.cc DEPS core)
@@ -47,6 +48,7 @@ cinn_cc_test(test_intrinsic_ops SRCS intrinsic_ops_test.cc DEPS cinncore)
 cinn_cc_test(test_ir_verify SRCS ir_verify_test.cc DEPS cinncore)
 cinn_cc_test(test_schedule_desc SRCS schedule_desc_test.cc DEPS cinncore)
 cinn_cc_test(test_ir_compare SRCS ir_compare_test.cc DEPS cinncore)
+cinn_cc_test(test_ir_copy SRCS ir_copy_test.cc DEPS cinncore)
 
 foreach(header ${schedule_desc_proto_HDRS})
   set(core_proto_includes
