+++ paddle/cinn/ir/ir_copy_test.cc
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "paddle/cinn/optim/ir_copy.h"
+#include "paddle/cinn/ir/ir_copy.h"
 
 #include <gtest/gtest.h>
