+++ paddle/cinn/optim/CMakeLists.txt
@@ -5,7 +5,6 @@ gather_srcs(
   SRCS
   remove_nested_block.cc
   replace_call_with_expr.cc
-  ir_copy.cc
   ir_replace.cc
   replace_var_with_expr.cc
   tensor_write_tell.cc
@@ -38,7 +37,6 @@ endif()
 
 cinn_cc_test(test_remove_nested_block SRCS remove_nested_block_test.cc DEPS
              cinncore)
-cinn_cc_test(test_ir_copy SRCS ir_copy_test.cc DEPS cinncore)
 cinn_cc_test(test_ir_simplify SRCS ir_simplify_test.cc DEPS cinncore)
 cinn_cc_test(test_replace_call_with_expr SRCS replace_call_with_expr_test.cc
              DEPS cinncore)
