+++ paddle/cinn/optim/ir_replace.cc
@@ -16,9 +16,9 @@
 
 #include <set>
 
+#include "paddle/cinn/ir/ir_copy.h"
 #include "paddle/cinn/ir/ir_mutator.h"
 #include "paddle/cinn/ir/ir_printer.h"
-#include "paddle/cinn/optim/ir_copy.h"
 #include "paddle/cinn/utils/string.h"
 
 namespace cinn {
