+++ paddle/cinn/optim/optimize.cc
@@ -14,6 +14,7 @@
 
 #include "paddle/cinn/optim/optimize.h"
 
+#include "paddle/cinn/ir/ir_copy.h"
 #include "paddle/cinn/ir/ir_printer.h"
 #include "paddle/cinn/ir/ir_schedule_util.h"
 #include "paddle/cinn/optim/call_arg_list_to_pod_value.h"
@@ -24,7 +25,6 @@
 #include "paddle/cinn/optim/fold_cinn_call_arguments.h"
 #include "paddle/cinn/optim/if_simplify.h"
 #include "paddle/cinn/optim/insert_debug_log_callee.h"
-#include "paddle/cinn/optim/ir_copy.h"
 #include "paddle/cinn/optim/ir_simplify.h"
 #include "paddle/cinn/optim/lower_function_call_bind_vars.h"
 #include "paddle/cinn/optim/lower_intrin.h"
