+++ paddle/cinn/optim/unroll_loops.cc
@@ -17,10 +17,10 @@
 #include <utility>
 #include <vector>
 
+#include "paddle/cinn/ir/ir_copy.h"
 #include "paddle/cinn/ir/ir_mutator.h"
 #include "paddle/cinn/ir/ir_operators.h"
 #include "paddle/cinn/ir/ir_printer.h"
-#include "paddle/cinn/optim/ir_copy.h"
 #include "paddle/cinn/optim/ir_replace.h"
 
 namespace cinn {
