+++ paddle/cinn/pybind/optim.cc
@@ -17,8 +17,8 @@
 #include "paddle/cinn/common/shared.h"
 #include "paddle/cinn/common/target.h"
 #include "paddle/cinn/common/type.h"
+#include "paddle/cinn/ir/ir_copy.h"
 #include "paddle/cinn/ir/ir_operators.h"
-#include "paddle/cinn/optim/ir_copy.h"
 #include "paddle/cinn/optim/ir_simplify.h"
 #include "paddle/cinn/pybind/bind.h"
 #include "paddle/cinn/pybind/bind_utils.h"
